Rename history_source_enum.py to history_source.py
